fix: sapphire-localnet: Wait for postgres and the KM to be ready
Postgres sometimes needs a bit more time to start up, which makes
the gateway fail.  We now wait for it before starting the gateway.

Also, the key manager sometimes needs longer to generate the public
ephemeral key needed for confidential contracts to work.  We now
wait for the epehemeral key to become available before continuing
with the rest of the container initialization.